Create Day 9 Dictionaries and Lists
Contains format to add nested dictionaries and lists. The travel_log list contains multiple dictionaries. A new dictionary is added with the append function. Format for adding to nested items on line 21. 

The "key" in python is not a variable, it is a string that is typed and can be changed whenever. Ie: when adding under the dictionary "country" it's not a list named country it's just a word you have to type in every time.